fix: Endpoints werent returning data
This happened because both /all and /individuals endpoints were meshed into one, and that was causing some data returning issues, due to over-engeneering.
