+++ packages/solidarity-client/server/utils/index.ts
@@ -13,3 +13,28 @@ export const getSupportType = subject => {
   const match = removeSpecialCaracters.match(/\b(psicolgico|jurdico|psicloga|advogada)\b/g)
   return match && match.length > 0 ? match : []
 }
+
+export const getTicketType = (type, subject) => {
+  if (type === 'psicológico_e_jurídico') {
+    const match = getSupportType(subject)
+    if (match.length > 0) {
+      return match[0] === 'jurdico'
+        ? 'jurídico'
+        : 'psicológico'
+    }
+  }
+  return typeof type !== 'undefined' ? type : '-'
+}
+
+export const fuseTicketsAndUsers = (users, tickets) => tickets
+  .filter(ticket => isValidTicket(users, ticket))
+  .map(ticket => {
+    const user = getUserFromTicket(users, ticket)[0]
+    return {
+      ...ticket,
+      ...user,
+      ticket_status: ticket.status,
+      ticket_created_at: ticket.created_at,
+      tipo_de_acolhimento: getTicketType(user.tipo_de_acolhimento, ticket.subject)
+    }
+  })
